Add development tools setting
This setting allows for the use of built-in development tools. Previously there was no way to use the development tools without compiling with the DEV variable set to true in main.js or test_main.js, after which you were stuck with developer tools. This allows you to toggle developer tools at will.